+++ lib/Logger.ts
@@ -37,6 +37,7 @@ export class Logger<
       short = false,
       levelLog = IOLevelLog.NONE,
       useColor = true,
+      transportType = "none",
     }: IOLoggerInterface,
     onInit: IOOnloadInterface<P> = (Logger) => {}
   ) {
@@ -48,6 +49,7 @@ export class Logger<
         short,
         levelLog,
         useColor,
+        transportType,
       },
       {} as P
     );
@@ -61,6 +63,7 @@ export class Logger<
       short = false,
       levelLog = IOLevelLog.NONE,
       useColor = true,
+      transportType = "none",
     }: IOLoggerInterface,
     onInit: IOOnloadInterface<P> = (Logger) => {}
   ) {
@@ -72,6 +75,7 @@ export class Logger<
         short,
         levelLog,
         useColor,
+        transportType,
       },
       onInit
     );